+++ pop/__main__.py
@@ -1,6 +1,5 @@
 # -*- coding: UTF-8 -*-
 
-#import argparse
 import os
 import sys
 import argparse
@@ -28,12 +27,21 @@ args = aparser.parse_args()
 if not args.input:
     exit()
 
+# Setup the build directory if there's one!
+if args.d:
+    os.chdir(args.d)
+
+# Save the builds path and set the compiler path!
+build_dir = os.getcwd()
+pop_dir = "\\".join(sys.argv[0].split("\\")[:-2])
+os.chdir(pop_dir)
+
 # We need an output file name!
 if not args.o:
     args.o = args.input[0].split(".")[0]
 
 # List of object files to compile together.
-objlist = ["prt.obj"]  # By default prt.obj is included.
+objlist = ["\"{0}\\prt.obj\"".format(build_dir)]  # By default prt.obj is included.
 
 for popfile in args.input:
     # Only compile pop source files!
@@ -41,29 +49,41 @@ for popfile in args.input:
         continue
 
     # Open the input file(s). NOTE: It takes only one file as of right now.
-    with open(popfile) as file:
+    with open("{0}\{1}".format(build_dir, popfile)) as file:
         data = file.read().split("\n")
     for i in range(len(data)):
         data[i] = data[i].rstrip()
+
+    # Change the working directory.
+    os.chdir(build_dir)
 
     # If everything went good, start the compiling proccess.
     a, b, c = Lexer().lexer(data, popfile)
     l = Parser().parser(a, b, c)
 
+    # Change the working directory.
+    os.chdir(pop_dir)
+
     if l:
-        with open("{0}.asm".format(popfile[:-4]), "w") as file:
+        with open("{0}\{1}.asm".format(build_dir, popfile[:-4]), "w") as file:
             file.write("\n".join(l))
 
         # Assemble the newly produced code.
-        os.system("nasm -f win32 -o {0}.obj {0}.asm".format(popfile[:-4]))
-        objlist.append("{0}.obj".format(popfile[:-4]))
+        os.system("nasm -f win32 -o \"{0}\\{1}.obj\" \"{0}\\{1}.asm\"".format(build_dir, popfile[:-4]))
+        objlist.append("\"{0}\\{1}.obj\"".format(build_dir, popfile[:-4]))
+
+# Change the working directory.
+os.chdir(pop_dir)
 
 # Assemble the pop runtime.
-os.system("nasm -f win32 -o prt.obj asm/prt.asm")
+os.system("nasm -f win32 -o \"{0}\prt.obj\" asm\\prt.asm".format(build_dir, pop_dir))
 
 # Link the object files together.
-os.system("Golink /console /entry __start /fo test.exe {0} kernel32.dll user32.dll msvcrt.dll".format(" ".join(objlist)))
+os.system("Golink /console /entry __start /fo \"{1}\\test.exe\" {0} kernel32.dll user32.dll msvcrt.dll".format(" ".join(objlist), build_dir))
+
+# Change the working directory.
+os.chdir(build_dir)
 
 # Clean up any leftover object files.
 for i in objlist:
-    os.remove(i)
+    os.remove(i[1:-1])