+++ test/shell.py
@@ -1,0 +1,7 @@
+def main():
+    print("Local OS shell via Python.")
+    while True:
+        i = input(">>> ")
+
+if __name__ == "__main__":
+    main()